require admin to delete

+++ client/src/components/MainPost.vue
@@ -5,23 +5,29 @@ import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import type { Post } from '@/types';
 import { faTrash } from '@fortawesome/free-solid-svg-icons';
 import { usePostStore } from '@/stores/post';
+import { useSessionStore } from '@/stores/session';
 
 const props = defineProps<{
     post?: Post
 }>()
 
 const postStore = usePostStore();
+const sessionStore = useSessionStore();
 
 function deletePost() {
-    postStore.deletePost(props.post?.id!);
+    if (sessionStore.user.admin) {
+        postStore.deletePost(props.post?.id!);
+    }
 }
 </script>
 
 <template>
     <div class="post">
         <div class="postinner"></div>
-        <div class="author cursor-pointer" @click="deletePost()">{{ post?.user }}
-            <FontAwesomeIcon :icon="faTrash" />
+        <div class="author">
+            {{ post?.user }}
+            <FontAwesomeIcon :icon="faTrash" :class="{ 'cursor-pointer': sessionStore.user.admin }"
+                @click="deletePost()" v-if="sessionStore.user.admin" />
         </div>
         <div class="title m-4">{{ post?.title }}</div>
         <div class="body">{{ post?.body }}</div>
